+++ CHANGELOG.md
@@ -27,14 +27,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Updated to Miniconda 24.4.0-0.
 - Use MacOSX matplotlib backend on macOS, still use TkAgg on Linux
+  - Update sed command to make this change
 
 ### Added
 
 - Explicit Conda Packages
   - earthaccess
-
 - Explicit Pip Packages
   - pyephem
+
+### Removed
+
+- Removed old sed command on `numpy` as file being altered no longer exists
 
 ## [24.1.2] - 2024-04-09